+++ CHANGELOG.md
@@ -1,24 +1,37 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.0.0] - 2019-09-24
+
+### Added
+- {{strtotime}} inserttag
+
+### Changed
+- refactoring
+
 ## [0.3.0] - 2019-08-23
 
-- added amp support for download inserttag
+### Added
+- amp support for download inserttag
 
 ## [0.2.0] - 2019-08-22
 
-- added small and endsmall insert tags
+### Added
+- small and endsmall insert tags
 - some smaller refactoring
 
 ## [0.1.2] - 2018-08-02
 
-- fixed download tag css class and id parameter not working
+### Fixed
+- download tag css class and id parameter not working
 
 ## [0.1.1] - 2018-08-02
 
-- fixed missing type for content download element
-- updated documentation
+### Fixed
+- missing type for content download element
+- documentation
 
 ## [0.1.0] - 2018-07-31
 
-Initial commit
+### Added
+- Initial commit